feat: custom user agent InternalCognitoUser (#11709)
* chore: port code from CognitoUser and CognitoUserPool to internals

* chore: delete InternalCognitoUserPool - separate PR

* chore: update imports and class name

* feat: add userAgentValue param to cognito client api's

* feat: add userAgentValue param to InternalCognitoUser api's, pass to client

* feat: make CognitoUser extend InternalCognitoUser and override api's

* test: update tests to work with additional param, minor fix

* chore: clean up unused imports and const's

* feat: add missing userAgentValue param

* chore: add new param to documentation comments

* build: export InternalCognitoUser from internals scope

* fix: remove default from InternalCognitoUser to fix scoped export

* fix: fix typo causing signOut failure

* build: correct index.d.ts files with changes

* test: revert auth tests

* chore: increase bundle size limits

* chore: simplify InternalCognitoUser import

---------
