+++ packages/amazon-cognito-identity-js/__tests__/internalsIndex.test.js
@@ -6,6 +6,7 @@ describe('import * keys', () => {
 		Array [
 		  "addAuthCategoryToCognitoUserAgent",
 		  "addFrameworkToCognitoUserAgent",
+		  "InternalCognitoUser",
 		]
 	`);
 	});
