+++ packages/amazon-cognito-identity-js/src/internals/index.js
@@ -2,3 +2,5 @@ export {
 	addAuthCategoryToCognitoUserAgent,
 	addFrameworkToCognitoUserAgent,
 } from '../UserAgent';
+
+export { InternalCognitoUser } from './InternalCognitoUser';
