+++ packages/api-graphql/package.json
@@ -65,7 +65,7 @@
       "name": "API (GraphQL client)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, GraphQLAPI }",
-      "limit": "88.35 kB"
+      "limit": "88.92 kB"
     }
   ],
   "jest": {
