+++ packages/api/package.json
@@ -67,7 +67,7 @@
 			"name": "API (top-level class)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, API }",
-			"limit": "89.07 kB"
+			"limit": "89.68 kB"
 		}
 	],
 	"jest": {
