+++ packages/auth/package.json
@@ -59,7 +59,7 @@
       "name": "Auth (top-level class)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, Auth }",
-      "limit": "55.15 kB"
+      "limit": "55.73 kB"
     }
   ],
   "jest": {
